requirements.txt
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.19.0
pillow==7.2.0
tensorflow==2.2.0

